[#510] Address pr feedback
